fix: use np.int64 insread of removed np.int (ref. https://numpy.org/devdocs/release/1.18.5-notes.html)

+++ 01compute_features/test_cepstrum.py
@@ -50,7 +50,7 @@ if __name__ == "__main__":
         waveform = np.frombuffer(waveform, dtype=np.int16)
 
     # 分析する時刻をサンプル番号に変換
-    target_index = np.int(target_time * sampling_frequency)
+    target_index = np.int64(target_time * sampling_frequency)
 
     # FFTを実施する区間分の波形データを取り出す
     frame = waveform[target_index:
